requirements.txt
# Django and Core Dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
django-filter==23.3
dj-database-url==2.1.0
psycopg2-binary==2.9.9

# QR Code and Image Processing
qrcode==7.4.2
# Use a newer Pillow version that has wheels for recent Python versions
Pillow>=10.4.0

# PDF Generation
reportlab==4.0.4
setuptools>=68.0.0
openpyxl>=3.1.2

# Async Task Processing
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Database
# SQLite is included with Python (no extra driver needed)
# For production PostgreSQL (commented out for local dev on Windows to avoid build tools requirement):
# psycopg2-binary==2.9.9

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Environment Variables
python-dotenv==1.0.0

# CORS and Security
django-environ==0.11.2

# Development Dependencies (optional)
# django-debug-toolbar==4.2.0
# django-extensions==3.2.3